refactor: Contributionを取得する処理を切り出した
