+++ server.ts
@@ -1,54 +1,5 @@
 import { createCanvas, serve } from "./deps.ts";
-
-interface ContributionDay {
-  color: string;
-  contributionCount: number;
-  contributionLevel: string;
-  date: string;
-}
-
-interface Week {
-  contributionDays: ContributionDay[];
-}
-
-interface ContributionCalendar {
-  totalContributions: number;
-  weeks: Week[];
-}
-
-const getContributions = async (user: string) => {
-  const token = Deno.env.get("GH_READ_USER_TOKEN");
-  const query = `
-    query($user:String!) {
-      user(login: $user){
-        contributionsCollection {
-          contributionCalendar {
-            totalContributions
-            weeks {
-              contributionDays {
-                color
-                contributionCount
-                contributionLevel
-                date
-              }
-            }
-          }
-        }
-      }
-    }
-  `;
-
-  const variables = { user };
-  const json = { query, variables };
-  const url = "https://api.github.com/graphql";
-  const res = await fetch(url, {
-    method: "POST",
-    headers: { Authorization: `Bearer ${token}` },
-    body: JSON.stringify(json),
-  });
-
-  return res.json();
-};
+import { ContributionDay, getContributions, Week } from "./contributions.ts";
 
 const port = 8080;
 
@@ -56,9 +7,6 @@ const handler = async (request: Request): Promise<Response> => {
   const url = new URL(request.url);
   const user = url.searchParams.get("user");
 
-  // const decoder = new TextDecoder("utf-8");
-  // const file = await Deno.readFile("contributions.json");
-  // const data = JSON.parse(decoder.decode(file));
   const data = await getContributions(user!);
 
   const canvas = createCanvas(670, 140);
